% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_smallbox_caption_positions.R
\name{check_smallbox_caption_positions}
\alias{check_smallbox_caption_positions}
\title{Check smallboxes intruding on chapter line}
\usage{
check_smallbox_caption_positions(path = ".", .report_error)
}
\arguments{
\item{path}{A path to a report that has been compiled.}

\item{.report_error}{Function to produce error messages.}
}
\value{
Called for its side-effect: errors in the unusual case when a smallbox of a particular size
}
\description{
Check smallboxes intruding on chapter line
}
