% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGrattanReport.R
\name{checkGrattanReport}
\alias{checkGrattanReport}
\alias{checkGrattanReports}
\title{Check Grattan Report}
\usage{
checkGrattanReport(path = ".", compile = FALSE, pre_release = FALSE,
  release = FALSE, .proceed_after_rerun, .no_log = TRUE,
  .log_psv = FALSE, embed = TRUE, rstudio = FALSE,
  update_grattan.cls = pre_release, filename = NULL,
  bib_warn_only = FALSE)

checkGrattanReports(path = ".", compile = FALSE, pre_release = FALSE,
  release = FALSE, .proceed_after_rerun, .no_log = TRUE,
  embed = TRUE, rstudio = FALSE, update_grattan.cls = TRUE)
}
\arguments{
\item{path}{Path to search for the tex source file.}

\item{compile}{Should \code{pdflatex} be run on the report so the logs may be checked?}

\item{pre_release}{Should the document be assumed to be final? Runs additional checks.}

\item{release}{Should a final pdf be prepared for publication?}

\item{.proceed_after_rerun}{On the occasions where infinitely many passes of \code{pdflatex}
are required, include this to skip the error. Note that this will result in false cross-references
or incompletely formatted bibliographies.}

\item{.no_log}{Make no entry in the log file on the check's outcome. (Deprecated.)}

\item{.log_psv}{If \code{TRUE}, return a table of the error details as pipe separated table.}

\item{embed}{If \code{FALSE}, not attempt to embed the fonts using Ghostscript is attempted. Useful if Ghostscript cannot easily be installed.
Set to \code{TRUE} for debugging or repetitive use (as in benchmarking).}

\item{rstudio}{Use the RStudio API if available.}

\item{update_grattan.cls}{Download \code{grattan.cls} from \url{https://github.com/grattan/grattex/blob/master/grattan.cls}?
Set to \code{FALSE} when checking the \code{grattex} repo itself. Also downloads the logos associated with the repository.}

\item{filename}{If provided, the \code{.tex} file inside \code{path} to check. By default, \code{NULL} so the
set to the sole \code{.tex} file within \code{path}.}

\item{bib_warn_only}{IF \code{TRUE}, show biber warnings as messages rather than errors.}
}
\value{
Called for its side-effect.
}
\description{
Check Grattan reports written is LaTeX for typing errors, significant warnings,
and inconsistent style.
}
\details{
Options in the package
\describe{
\item{\code{grattanReporter.quiet.progress = FALSE}}{Can be set to \code{TRUE} to avoid printing the output during the check.}
}
}
