% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_FrontPage.R
\name{compress_FrontPage}
\alias{compress_FrontPage}
\title{Compress the FrontPage}
\usage{
compress_FrontPage(path = ".", error = FALSE, msg.if.error = TRUE)
}
\arguments{
\item{path}{The path of a report, under which is a file \code{FrontPage/FrontPage.pdf}.}

\item{error}{If the compression does not succeed (for what ever reason), should an error be given? 
By default, \code{FALSE}.}

\item{msg.if.error}{Should a message be printed at all, if \code{error} is \code{FALSE}? If \code{TRUE}, the default,
the error message is printed via \code{message}, not \code{stop}, so allows subsequent code to continue. Otherwise, 
provided \code{error = FALSE}, the function returns \code{NULL}, invisibly.}
}
\value{
Called for its side-effect: attempts to compress \code{./FrontPage/FrontPage.pdf} using Ghostscript, in order to 
make the front page prettier than normal, but still a reasonable size.

Returns \code{NULL}, invisibly.
}
\description{
Compress the FrontPage
}
