% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctly_spelled_words_fixed_txt.R
\docType{data}
\name{correctly_spelled_words_fixed_txt}
\alias{correctly_spelled_words_fixed_txt}
\title{List of correctly spelled words}
\format{A character vector of words as fixed words to skip during the spell check.}
\usage{
correctly_spelled_words_fixed_txt
}
\description{
List of correctly spelled words
}
\keyword{datasets}
