% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_labels.R
\name{fix_labels}
\alias{fix_labels}
\title{Fix badly-prefixed labels}
\usage{
fix_labels(filename, out.file = filename)
}
\arguments{
\item{filename}{The filename whose contents is to be fixed.}

\item{out.file}{The filename where the fixed version of \code{filename} is to be written. By default, \code{filename}; i.e. \code{filename} is overwritten by this process.}
}
\description{
Fix badly-prefixed labels
}
