% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORRECTLY_SPELLED_WORDS_CASE_SENSITIVE.R
\docType{data}
\name{grattan_CORRECTLY_SPELLED_WORDS_CASE_SENSITIVE}
\alias{grattan_CORRECTLY_SPELLED_WORDS_CASE_SENSITIVE}
\title{List of correctly spelled, case-sensitive words}
\format{A character vector of words as perl-regex case-sensitive patterns to skip during the spell check.}
\usage{
grattan_CORRECTLY_SPELLED_WORDS_CASE_SENSITIVE
}
\description{
List of correctly spelled, case-sensitive words
}
\keyword{datasets}
