% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_travis.R
\name{setup_travis}
\alias{setup_travis}
\title{Setup Travis-CI for grattan reports}
\usage{
setup_travis()
}
\description{
Should not be invoked by end-users. Only to be run within the \code{.travis.yml} file
while in travis
}
