% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_bibliography.R
\name{validate_bibliography}
\alias{validate_bibliography}
\title{Validate bibliography according Grattan style}
\usage{
validate_bibliography(
  path = ".",
  file = NULL,
  .report_error,
  rstudio = FALSE,
  check_comma = TRUE
)
}
\arguments{
\item{path}{Containing the bib file.}

\item{file}{The bib file if specified.}

\item{.report_error}{How errors should be reported.}

\item{rstudio}{Use the RStudio API?}

\item{check_comma}{Logical. If `TRUE`, an error will be given if each .bib entry 
does not end with a comma. `TRUE` by default.}
}
\value{
\code{NULL} if bibliography validated.
}
\description{
Validate bibliography according Grattan style
}
