% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2tex.R
\name{word2tex}
\alias{word2tex}
\title{Convert word to tex}
\usage{
word2tex(path = ".", overwrite = TRUE)
}
\arguments{
\item{path}{Path containing the \code{.docx} file.}

\item{overwrite}{(logical, default: \code{TRUE}) Overwrite existing tex files.}
}
\description{
Convert word to tex
}
